ai/agents/task_agent/requirements.txt
google-adk
a2a-sdk[all]
litellm
python-dotenv
